Order created and working

+++ database/migrations/2024_05_15_120300_create_pedidos_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             
             $table->date('fecha');
             $table->integer('unidades')->nullable();
-            $table->bigInteger('coste')->nullable();
+            $table->decimal('coste')->nullable();
             $table->string('recordatorio');
 
             $table->timestamps();
